+++ src/app/profile/index.js
@@ -7,15 +7,21 @@ import ProfileCard from '../../components/profile-card';
 import LoginMenu from '../../components/login-menu';
 import useTranslate from '../../hooks/use-translate';
 import useSelector from '../../hooks/use-selector';
+import useStore from '../../hooks/use-store';
+import useInit from '../../hooks/use-init';
 
 function Profile() {
   const { t } = useTranslate();
+  const store = useStore();
+
+  useInit(() => {
+    store.actions.user.getUser();
+  }, []);
 
   const select = useSelector(state => ({
     name: state.user.name,
     phone: state.user.phone,
     email: state.user.email,
-    successfully: state.user.successfully,
   }));
 
   return (
