+++ src/store/exports.js
@@ -5,3 +5,4 @@ export { default as article } from './article';
 export { default as locale } from './locale';
 export { default as categories } from './categories';
 export { default as user } from './user';
+export { default as session } from './session';
